requirements.txt
astroid==2.5.3
azure-common==1.1.27
azure-functions==1.7.0
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycparser==2.20
pylint==2.7.4
pyodbc==4.0.30
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
typed-ast==1.4.3
urllib3==1.26.4
wrapt==1.12.1
